+++ tas/backend/helpers/config.py
@@ -41,12 +41,15 @@ config = {
     }
 }
 
-if os.path.isfile('config.json'):
-    with open('config.json', 'r') as f:
-        config.update(json.loads(f.read()))
-else:
-    with open('config.json', 'w') as f:
-        f.write(json.dumps(config, indent=4))
+
+def reload_config():
+    global config
+    if os.path.isfile('config.json'):
+        with open('config.json', 'r') as f:
+            config.update(json.load(f))
+    else:
+        with open('config.json', 'w') as f:
+            f.write(json.dumps(config, indent=4))
 
 
 def get_config(portion=None):
@@ -65,3 +68,6 @@ def set_config(nconfig, portion=None):
         config[portion] = nconfig
     with open('config.json', 'w') as f:
         f.write(json.dumps(config, indent=4))
+
+
+reload_config()